- fix [inexplicable] test issues by applying UI defaults in AuthenticationController   as opposed to ServerConfig. This also protects against multiple calls to setServerConfig   applying the defaults multiple times. - allow 'pause' to be case insensitive when set as the refresh default. - fix a console message from .log to .debug
