+++ src/app/AuthenticationController.tsx
@@ -14,7 +14,7 @@ import { ServerStatus } from '../types/ServerStatus';
 import InitializingScreen from './InitializingScreen';
 import { isKioskMode } from '../utils/SearchParamUtils';
 import * as AlertUtils from '../utils/AlertUtils';
-import { setServerConfig } from '../config/ServerConfig';
+import { setServerConfig, serverConfig, humanDurations } from '../config/ServerConfig';
 import { TLSStatus } from '../types/TLSStatus';
 import { MeshTlsActions } from '../actions/MeshTlsActions';
 import { AuthStrategy } from '../types/Auth';
@@ -23,16 +23,23 @@ import { LoginActions } from '../actions/LoginActions';
 import history from './History';
 import { NamespaceActions } from 'actions/NamespaceAction';
 import Namespace from 'types/Namespace';
+import { UserSettingsActions } from 'actions/UserSettingsActions';
+import { DurationInSeconds, IntervalInMilliseconds } from 'types/Common';
+import { config } from 'config';
+import { store } from 'store/ConfigStore';
 
 interface AuthenticationControllerReduxProps {
   authenticated: boolean;
   checkCredentials: () => void;
   isLoginError: boolean;
   landingRoute?: string;
+  setActiveNamespaces: (namespaces: Namespace[]) => void;
+  setDuration: (duration: DurationInSeconds) => void;
   setJaegerInfo: (jaegerInfo: JaegerInfo | null) => void;
   setLandingRoute: (route: string | undefined) => void;
   setMeshTlsStatus: (meshStatus: TLSStatus) => void;
   setNamespaces: (namespaces: Namespace[], receivedAt: Date) => void;
+  setRefreshInterval: (interval: IntervalInMilliseconds) => void;
   setServerStatus: (serverStatus: ServerStatus) => void;
 }
 
@@ -175,6 +182,7 @@ class AuthenticationController extends React.Component<AuthenticationControllerP
       ]);
       this.props.setNamespaces(configs[0].data, new Date());
       setServerConfig(configs[1].data);
+      this.applyUIDefaults();
 
       if (this.props.landingRoute) {
         history.replace(this.props.landingRoute);
@@ -186,6 +194,68 @@ class AuthenticationController extends React.Component<AuthenticationControllerP
       this.setState({ isPostLoginError: true });
     }
   };
+
+  private applyUIDefaults() {
+    const uiDefaults = serverConfig.kialiFeatureFlags.uiDefaults;
+    if (uiDefaults) {
+      // Duration (aka metricsPerRefresh)
+      if (uiDefaults.metricsPerRefresh) {
+        const validDurations = humanDurations(serverConfig, '', '');
+        let metricsPerRefresh = 0;
+        for (const [key, value] of Object.entries(validDurations)) {
+          if (value === uiDefaults.metricsPerRefresh) {
+            metricsPerRefresh = Number(key);
+            break;
+          }
+        }
+        if (metricsPerRefresh > 0) {
+          this.props.setDuration(metricsPerRefresh);
+          console.debug(
+            `Setting UI Default: metricsPerRefresh [${uiDefaults.metricsPerRefresh}=${metricsPerRefresh}s]`
+          );
+        } else {
+          console.debug(`Ignoring invalid UI Default: metricsPerRefresh [${uiDefaults.metricsPerRefresh}]`);
+        }
+      }
+
+      // Refresh Interval
+      let refreshInterval = -1;
+      if (uiDefaults.refreshInterval) {
+        for (const [key, value] of Object.entries(config.toolbar.refreshInterval)) {
+          if (value.toLowerCase().endsWith(uiDefaults.refreshInterval.toLowerCase())) {
+            refreshInterval = Number(key);
+            break;
+          }
+        }
+        if (refreshInterval >= 0) {
+          this.props.setRefreshInterval(refreshInterval);
+          console.debug(`Setting UI Default: refreshInterval [${uiDefaults.refreshInterval}=${refreshInterval}ms]`);
+        } else {
+          console.debug(`Ignoring invalid UI Default: refreshInterval [${uiDefaults.refreshInterval}]`);
+        }
+      }
+
+      // Selected Namespaces
+      if (uiDefaults.namespaces && uiDefaults.namespaces.length > 0) {
+        // use store directly, we don't want to update on redux state change
+        const namespaces = store.getState().namespaces.items;
+        const namespaceNames: string[] = namespaces ? namespaces.map(ns => ns.name) : [];
+        const activeNamespaces: Namespace[] = [];
+
+        for (const name of uiDefaults.namespaces) {
+          if (namespaceNames.includes(name)) {
+            activeNamespaces.push({ name: name } as Namespace);
+          } else {
+            console.debug(`Ignoring invalid UI Default: namespace [${name}]`);
+          }
+        }
+        if (activeNamespaces.length > 0) {
+          this.props.setActiveNamespaces(activeNamespaces);
+          console.debug(`Setting UI Default: namespaces ${JSON.stringify(activeNamespaces.map(ns => ns.name))}`);
+        }
+      }
+    }
+  }
 
   private setDocLayout = () => {
     if (document.documentElement) {
@@ -212,10 +282,13 @@ const mapStateToProps = (state: KialiAppState) => ({
 
 const mapDispatchToProps = (dispatch: KialiDispatch) => ({
   checkCredentials: () => dispatch(LoginThunkActions.checkCredentials()),
+  setActiveNamespaces: bindActionCreators(NamespaceActions.setActiveNamespaces, dispatch),
+  setDuration: bindActionCreators(UserSettingsActions.setDuration, dispatch),
   setJaegerInfo: bindActionCreators(JaegerActions.setInfo, dispatch),
   setLandingRoute: bindActionCreators(LoginActions.setLandingRoute, dispatch),
   setMeshTlsStatus: bindActionCreators(MeshTlsActions.setinfo, dispatch),
   setNamespaces: bindActionCreators(NamespaceActions.receiveList, dispatch),
+  setRefreshInterval: bindActionCreators(UserSettingsActions.setRefreshInterval, dispatch),
   setServerStatus: (serverStatus: ServerStatus) => processServerStatus(dispatch, serverStatus)
 });
