+++ src/config/ServerConfig.ts
@@ -1,12 +1,6 @@
 import _ from 'lodash';
-
 import { ServerConfig } from '../types/ServerConfig';
 import { parseHealthConfig } from './HealthConfig';
-import { store } from 'store/ConfigStore';
-import { UserSettingsActions } from 'actions/UserSettingsActions';
-import { config } from 'config';
-import Namespace from 'types/Namespace';
-import { NamespaceActions } from 'actions/NamespaceAction';
 
 export type Durations = { [key: number]: string };
 
@@ -14,8 +8,8 @@ export type ComputedServerConfig = ServerConfig & {
   durations: Durations;
 };
 
-export const humanDurations = (config: ComputedServerConfig, prefix?: string, suffix?: string) =>
-  _.mapValues(config.durations, v => _.reject([prefix, v, suffix], _.isEmpty).join(' '));
+export const humanDurations = (cfg: ComputedServerConfig, prefix?: string, suffix?: string) =>
+  _.mapValues(cfg.durations, v => _.reject([prefix, v, suffix], _.isEmpty).join(' '));
 
 const toDurations = (tupleArray: [number, string][]): Durations => {
   const obj = {};
@@ -96,74 +90,14 @@ export const toValidDuration = (duration: number): number => {
   return durationsTuples[0][0];
 };
 
-export const setServerConfig = (svcConfig: ServerConfig) => {
+export const setServerConfig = (cfg: ServerConfig) => {
   serverConfig = {
-    ...svcConfig,
+    ...cfg,
     durations: {}
   };
-  serverConfig.healthConfig = svcConfig.healthConfig
-    ? parseHealthConfig(svcConfig.healthConfig)
-    : serverConfig.healthConfig;
+  serverConfig.healthConfig = cfg.healthConfig ? parseHealthConfig(cfg.healthConfig) : serverConfig.healthConfig;
 
   computeValidDurations(serverConfig);
-
-  // apply configured UI Defaults
-  const uiDefaults = serverConfig.kialiFeatureFlags.uiDefaults;
-  if (uiDefaults) {
-    // Duration (aka metricsPerRefresh)
-    if (uiDefaults.metricsPerRefresh) {
-      const validDurations = humanDurations(serverConfig, '', '');
-      let metricsPerRefresh = 0;
-      for (const [key, value] of Object.entries(validDurations)) {
-        if (value === uiDefaults.metricsPerRefresh) {
-          metricsPerRefresh = Number(key);
-          break;
-        }
-      }
-      if (metricsPerRefresh > 0) {
-        store.dispatch(UserSettingsActions.setDuration(metricsPerRefresh));
-        console.debug(`Setting UI Default: metricsPerRefresh [${uiDefaults.metricsPerRefresh}=${metricsPerRefresh}s]`);
-      } else {
-        console.debug(`Ignoring invalid UI Default: metricsPerRefresh [${uiDefaults.metricsPerRefresh}]`);
-      }
-    }
-
-    // Refresh Interval
-    let refreshInterval = -1;
-    if (uiDefaults.refreshInterval) {
-      for (const [key, value] of Object.entries(config.toolbar.refreshInterval)) {
-        if (value.endsWith(uiDefaults.refreshInterval)) {
-          refreshInterval = Number(key);
-          break;
-        }
-      }
-      if (refreshInterval >= 0) {
-        store.dispatch(UserSettingsActions.setRefreshInterval(refreshInterval));
-        console.debug(`Setting UI Default: refreshInterval [${uiDefaults.refreshInterval}=${refreshInterval}ms]`);
-      } else {
-        console.debug(`Ignoring invalid UI Default: refreshInterval [${uiDefaults.refreshInterval}]`);
-      }
-    }
-
-    // Selected Namespaces
-    if (uiDefaults.namespaces && uiDefaults.namespaces.length > 0) {
-      const namespaces = store.getState().namespaces.items;
-      const namespaceNames: string[] = namespaces ? namespaces.map(ns => ns.name) : [];
-      const activeNamespaces: Namespace[] = [];
-
-      for (const name of uiDefaults.namespaces) {
-        if (namespaceNames.includes(name)) {
-          activeNamespaces.push({ name: name } as Namespace);
-        } else {
-          console.debug(`Ignoring invalid UI Default: namespace [${name}]`);
-        }
-      }
-      if (activeNamespaces.length > 0) {
-        store.dispatch(NamespaceActions.setActiveNamespaces(activeNamespaces));
-        console.log(`Setting UI Default: namespaces ${JSON.stringify(activeNamespaces.map(ns => ns.name))}`);
-      }
-    }
-  }
 };
 
 export const isIstioNamespace = (namespace: string): boolean => {
